requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
openpyxl>=3.1.0
azure-identity>=1.15.0
requests>=2.31.0
tenacity>=8.2.0
plotly>=5.18.0
